# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API CAPPirote",
    "description": "Documentación de la API CAPPirote",
    "version": "1.0.2"
  },
  "servers": [
    {
      "url": "https://cappirote-bf5611b04e17.herokuapp.com/api"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/hermandades": {
      "get": {
        "tags": [
          "hermandades"
        ],
        "summary": "Obtener todas las hermandades",
        "description": "Recupera todas las hermandades disponibles en el sistema.",
        "operationId": "getAllHdads",
        "responses": {
          "200": {
            "description": "Lista de hermandades recuperada con éxito",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "ID de la hermandad",
                        "example": 1
                      },
                      "nombre": {
                        "type": "string",
                        "description": "Nombre de la hermandad",
                        "example": "Hdad de El Prado"
                      },
                      "apelativo": {
                        "type": "string",
                        "description": "Apelativo o nombre corto de la hermandad",
                        "example": "El Prado"
                      },
                      "fundacion": {
                        "type": "integer",
                        "description": "Año de fundación de la hermandad",
                        "example": 2009
                      },
                      "titulares": {
                        "type": "string",
                        "description": "Titulares de la hermandad",
                        "example": "Ntra Señora del Prado en su Dolor"
                      },
                      "sede": {
                        "type": "string",
                        "description": "Sede de la hermandad",
                        "example": "Capilla de Nuestra Señora del Prado en su Dolor"
                      },
                      "dia": {
                        "type": "string",
                        "description": "Día de la celebración de la hermandad",
                        "example": "Viernes de Dolores"
                      },
                      "itinerario": {
                        "type": "string",
                        "description": "Itinerario de la hermandad",
                        "example": "Templo, Plaza Fuentepiña, Costa Rica, Sainete, Escacena del Campo..."
                      },
                      "escudo": {
                        "type": "string",
                        "description": "Ruta de la imagen del escudo de la hermandad",
                        "example": "assets/images/Escudo-Prado.jpg"
                      },
                      "latitud": {
                        "type": "number",
                        "format": "float",
                        "description": "Latitud",
                        "example": null
                      },
                      "longitud": {
                        "type": "number",
                        "format": "float",
                        "description": "Longitud",
                        "example": null
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno en el servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Error interno del servidor al recuperar las hermandades."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "hermandades"
        ],
        "summary": "Crea una nueva hermandad",
        "description": "Crea una nueva hermandad",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "object",
                    "example": "Hdad de Prueba"
                  },
                  "apelativo": {
                    "type": "object",
                    "example": "La Prueba"
                  },
                  "titulares": {
                    "type": "object",
                    "example": "Santo Cristo Cautivo y María Santísima de la Misericordia"
                  },
                  "fundacion": {
                    "type": "object",
                    "example": "1981"
                  },
                  "sede": {
                    "type": "object",
                    "example": "Capilla del Santo Cristo Cautivo"
                  },
                  "dia": {
                    "type": "object",
                    "example": "Lunes Santo"
                  },
                  "itinerario": {
                    "type": "object",
                    "example": "Templo, Plaza de la Misericordia, Ruiz de Alarcón, Fernando de Villalón, Rubén Darío, Chucena, Virgen de Montemayor, Virgen de las Angustias, Avenida Santa Marta, Cristo de la Redención, Rotonda Huerto Paco, Avenida Adoratrices, Fray Junípero Serra, Fray Juan Pérez, San Sebastián, Plaza de Los Litris, Jesús de la Pasión, Plaza de San Pedro, Daoiz, Paseo Santa Fe, Puerto, Cardenal Albornoz, José Nogales, La Placeta, CARRERA OFICIAL, San Salvador, Fernando el Católico, Palos, Plaza Quintero Báez, La Fuente, Plaza de San Pedro, San Andrés, Plaza Ivonne Cazenave, Avenida Manuel Siurot, Avenida San Antonio, Rotonda Huerto Paco, Cristo de la Redención, Avenida Santa Marta, Virgen del Reposo, Manzanilla, Virgen de Montemayor, Remo, Díaz del Castillo, Pasaje del Cautivo, Plaza de la Misericordia y Templo."
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
  },
...
}